+++ CHANGELOG.md
@@ -1,5 +1,14 @@
+## 0.0.6
+* Made default values of baseColor and highlightColor and nullable for them.
+
 ## 0.0.5
-* Just update image url.
+* Add Images to README.md file.
+
+## 0.0.4
+* Update README.md file.
+
+## 0.0.3
+* Add Example.
 
 ## 0.0.2
 * Initial release.